cloudops/devops/requirements.txt
amqp==2.2.2
asgiref==2.2.0
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
bcrypt==3.1.4
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
cffi==1.11.4
channels==2.1.1
channels_redis==2.2.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
daphne==2.1.2
Django==2.0.1
django-celery-beat==1.1.1
django-core==1.4.1
django-mssql==1.8
django-tinymce==2.7.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
kombu==4.1.0
mysql-connector-python==8.0.5
mysqlclient==1.3.12
paramiko==2.4.0
Pillow==5.1.0
pyasn1==0.4.2
pycparser==2.18
pycryptodome==3.5.1
pymssql==2.1.3
PyMySQL==0.8.0
PyNaCl==1.2.1
pytz==2017.3
redis==2.10.6
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
Twisted==17.9.0
txaio==2.9.0
urllib3==1.22
vine==1.1.4
zope.interface==4.4.3
python-jenkins==1.0.0
xlrd==1.1.0
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
django-cors-headers==2.4.0
boto3==1.7.32
aliyun-python-sdk-core==2.3.5
aliyun-python-sdk-core-v3==2.8.6
aliyun-python-sdk-ecs==4.9.0
aliyun-python-sdk-rds==2.1.4
django-guardian==1.4.9
pandas==0.23.3
pymongo==3.7.1
pycurl==7.43.0.2
